+++ frontend/src/Settings/Profiles/Delay/DelayProfile.js
@@ -85,7 +85,7 @@ class DelayProfile extends Component {
       connectDragSource
     } = this.props;
 
-    let preferred = titleCase(preferredProtocol);
+    let preferred = `Prefer ${titleCase(preferredProtocol)}`;
 
     if (!enableUsenet) {
       preferred = translate('OnlyTorrent');
